Capstone3/requirements.txt
streamlit
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
py-trello
python-dotenv
nest_asyncio
